microservices/privacy-service/requirements.txt
flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.0.1
facenet-pytorch==2.5.3
redis==5.0.1
werkzeug==3.0.1
torch==2.1.0
torchvision==0.16.0
requests==2.31.0